% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_search.R
\name{add_word}
\alias{add_word}
\title{Add a word to a wordsearch matrix}
\usage{
add_word(x, word = "finding")
}
\arguments{
\item{x}{wordsearch matrix}

\item{word}{the word to add (character/scalar)}
}
\description{
Add a word to a wordsearch matrix
}
