% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{as_crossword}
\alias{as_crossword}
\title{Assign an object to the `crossword` class}
\usage{
as_crossword(x)
}
\arguments{
\item{x}{an object containing crossword data}
}
\description{
Assign an object to the `crossword` class
}
