% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{book}
\alias{book}
\title{Create a wordsearch/crossword book}
\usage{
book(
  input_file = system.file("book.yml", package = "worrrd"),
  output_file = "book.Rmd"
)
}
\arguments{
\item{input_file}{yaml file containing book details/contents}

\item{output_file}{full path to output file}
}
\description{
Create a wordsearch/crossword book
}
