% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_by_numbers.R
\name{color_by_numbers}
\alias{color_by_numbers}
\title{Convert an image to a color-by-numbers}
\usage{
color_by_numbers(img)
}
\arguments{
\item{img}{color image to convert}

\item{n_colors}{number of colors to include}
}
\description{
Convert an image to a color-by-numbers
}
