% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{is_crossword}
\alias{is_crossword}
\title{Check if an object is of the `crossword` class}
\usage{
is_crossword(x)
}
\arguments{
\item{x}{an R object to check}
}
\value{
logical/scalar
}
\description{
Check if an object is of the `crossword` class
}
