% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_search.R
\name{max_word_size}
\alias{max_word_size}
\title{Compute maximum word size, based on a word_search matrix}
\usage{
max_word_size(x)
}
\arguments{
\item{x}{word_search matrix}
}
\description{
Compute maximum word size, based on a word_search matrix
}
