% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsearch.R
\name{plot.wordsearch}
\alias{plot.wordsearch}
\title{Draw a wordsearch puzzle}
\usage{
\method{plot}{wordsearch}(x, solution = FALSE, letter_size = 8)
}
\arguments{
\item{x}{wordsearch object}

\item{solution}{show solution? (logical/scalar)}

\item{letter_size}{size of letters (numeric/scalar)}
}
\description{
Draw a wordsearch puzzle
}
