% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_search.R
\name{plot_word_search}
\alias{plot_word_search}
\title{Draw a wordsearch}
\usage{
plot_word_search(x, solution = FALSE, letter_size = 8)
}
\arguments{
\item{x}{wordsearch object}

\item{solution}{show solution? (logical/scalar)}

\item{letter_size}{size of letters (numeric/scalar)}
}
\description{
Draw a wordsearch
}
