% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{print.crossword}
\alias{print.crossword}
\title{Print a crossword puzzle}
\usage{
\method{print}{crossword}(x, ...)
}
\arguments{
\item{x}{a crossword object (see \code{\link{crossword}})}

\item{...}{additional printing args}
}
\value{
crossword object
}
\description{
Print a crossword puzzle
}
