% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsearch.R
\name{print.wordsearch}
\alias{print.wordsearch}
\title{Print details for a wordsearch puzzle}
\usage{
\method{print}{wordsearch}(x, ...)
}
\arguments{
\item{x}{wordsearch object (class: wordsearch)}

\item{...}{additional printing args}
}
\description{
Print details for a wordsearch puzzle
}
