% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printable.R
\name{printable}
\alias{printable}
\title{Prepare a gamer object for printing}
\usage{
printable(x, filename = "plot.pdf")
}
\arguments{
\item{x}{ggplot object}

\item{filename}{name of file}
}
\description{
Prepare a gamer object for printing
}
