% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_intersections.R, R/wordsearch.R
\name{word_intersections}
\alias{word_intersections}
\alias{word_intersections}
\title{Get possible intersection points based on the current board and a provided word}
\usage{
word_intersections(x, word = "needles")

word_intersections(x, word = "needles")
}
\arguments{
\item{x}{word matrix}

\item{word}{the word to add (character/scalar)}

\item{x}{wordsearch matrix}

\item{word}{the word to add (character/scalar)}
}
\description{
Get possible intersection points based on the current board and a provided word

Get possible intersection points based on the current board and a provided word
}
