% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_matrix.R
\name{word_matrix}
\alias{word_matrix}
\title{Create an empty word matrix}
\usage{
word_matrix(r = 10, c = 10)
}
\arguments{
\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Create an empty word matrix
}
