% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_overlap.R
\name{word_overlap}
\alias{word_overlap}
\title{Compute overlap score for a vector of words}
\usage{
word_overlap(words)
}
\arguments{
\item{words}{vector of words (character/vector)}
}
\description{
Compute overlap score for a vector of words
}
