% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_search.R
\name{word_search}
\alias{word_search}
\title{Create a word search}
\usage{
word_search(words = c("finding", "needles", "inside", "haystacks"), r = 10,
  c = 10)
}
\arguments{
\item{words}{a vector of hidden words (character/vector)}

\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Create a word search
}
