% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsearch.R
\name{wordsearch}
\alias{wordsearch}
\title{Create a wordsearch puzzle}
\usage{
wordsearch(words = c("finding", "needles", "inside", "haystacks"), r = 10,
  c = 10)
}
\arguments{
\item{words}{a vector of hidden words (character/vector)}

\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Create a wordsearch puzzle
}
