% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_word.R, R/wordsearch.R
\name{add_word}
\alias{add_word}
\alias{add_word}
\title{Add a word to a word matrix}
\usage{
add_word(x, word = "finding", must_intersect = FALSE)

add_word(x, word = "finding", must_intersect = FALSE)
}
\arguments{
\item{x}{word matrix}

\item{word}{the word to add (character/scalar)}

\item{must_intersect}{force the added word to intersect with >1 word (logical/scalar)}

\item{x}{wordsearch matrix}

\item{word}{the word to add (character/scalar)}
}
\description{
Add a word to a word matrix

Add a word to a wordsearch matrix
}
