% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{as_crossword}
\alias{as_crossword}
\title{Assign an object to the `crossword` class}
\usage{
as_crossword(x)
}
\arguments{
\item{x}{an object containing crossword data}
}
\value{
crossword object:
  a matrix reprepresentation of the crossword, with attributes:
     positions:  tibble representation of crossword
     clues:      tibble representation of clue start (i.e., clue number locations)
}
\description{
Assign an object to the `crossword` class
}
