% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsearch.R
\name{as_wordsearch}
\alias{as_wordsearch}
\title{Assign an object to the `wordsearch` class}
\usage{
as_wordsearch(x)
}
\arguments{
\item{x}{an object containing wordsearch data}
}
\description{
Assign an object to the `wordsearch` class
}
