% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.R
\name{book}
\alias{book}
\title{Create a puzzle book}
\usage{
book(
  input_file = system.file("book.yml", package = "worrrd"),
  output_file = "book.pdf",
  solutions = TRUE
)
}
\arguments{
\item{input_file}{yaml file containing book details/contents}

\item{output_file}{full path to output file (with .pdf extension)}

\item{solutions}{include solutions (logical/scalar)

#' @examples
\donttest{
# Create demo book included with package
book()
}}
}
\description{
Create a puzzle book
}
