% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{crossword}
\alias{crossword}
\title{Create a crossword puzzle}
\usage{
crossword(x, r = 50, c = 50)
}
\arguments{
\item{x}{a 2-column list/tibble/data.frame containing words and clues}

\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Create a crossword puzzle
}
