% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{crossword}
\alias{crossword}
\title{Create a crossword puzzle}
\usage{
crossword(words = c("finding", "needles", "inside", "haystacks"), r = 10,
  c = 10)
}
\arguments{
\item{words}{a vector of hidden words (character/vector)}

\item{r}{number of rows}

\item{c}{number of columns}
}
\description{
Create a crossword puzzle
}
