% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{crossword}
\alias{crossword}
\title{Create a crossword puzzle}
\usage{
crossword(words, clues, r = 50, c = 50)
}
\arguments{
\item{words}{a vector of words (character/vector)}

\item{clues}{a vector a clues (character/vector)}

\item{r}{number of rows (numeric/scalar)}

\item{c}{number of columns (numeric/scalar)}
}
\description{
Create a crossword puzzle
}
