% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_matrix.R
\name{image_matrix}
\alias{image_matrix}
\title{Convert an image to a 0/1 matrix}
\usage{
image_matrix(
  img = "https://upload.wikimedia.org/wikipedia/commons/9/96/Tux_Paint_banana.svg",
  r = 10,
  c = 10
)
}
\arguments{
\item{img}{image path}

\item{r}{number of rows (numeric/sclar)}
}
\description{
Convert an image to a 0/1 matrix
}
