% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsearch.R
\name{is_wordsearch}
\alias{is_wordsearch}
\title{Check if an object is of the `wordsearch` class}
\usage{
is_wordsearch(x)
}
\arguments{
\item{x}{an R object to check}
}
\description{
Check if an object is of the `wordsearch` class
}
