% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossword.R
\name{plot.crossword}
\alias{plot.crossword}
\title{Plot a crossword puzzle}
\usage{
\method{plot}{crossword}(
  x,
  solution = FALSE,
  clues = FALSE,
  title = "Crossword Puzzle",
  legend_size = 4,
  ...
)
}
\arguments{
\item{x}{a crossword object (see \code{\link{crossword}})}

\item{solution}{show solution? (logical/scalar)}

\item{clues}{show clues? (logical/scalar)}

\item{title}{puzzle title (character/scalar)}

\item{legend_size}{letter size of word list; set to NULL to auto-size (numeric/scalar)}

\item{...}{additional printing args}
}
\value{
ggplot2 object
}
\description{
Plot a crossword puzzle
}
