% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printable.R
\name{printable}
\alias{printable}
\title{Prepare a worrrd object for printing}
\usage{
printable(x, filename = "plot.pdf")
}
\arguments{
\item{x}{ggplot object}

\item{filename}{name of file}
}
\value{
filename of pdf puzzle
}
\description{
Prepare a worrrd object for printing
}
\examples{
words <- c("dog", "cat", "horse", "frog", "cow", "fox")
ex1 <- wordsearch(words, r = 10, c = 10)
my_puzzle <- plot(ex1, solution = FALSE)
printable(my_puzzle, "my_wordsearch.pdf")
unlink("my_wordsearch.pdf")

}
