% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_intersections.R
\name{word_intersections}
\alias{word_intersections}
\title{Get possible intersection points based on the current board and a provided word}
\usage{
word_intersections(x, word = "needles")
}
\arguments{
\item{x}{word matrix}

\item{word}{the word to add (character/scalar)}
}
\value{
for each direction, a matrix of crossing-point counts
}
\description{
Get possible intersection points based on the current board and a provided word
}
